requirements.txt
<<<<<<< HEAD
pyTelegramBotAPI
Flask
requests
python-dotenv
=======
pyTelegramBotAPI
Flask
requests
python-dotenv
>>>>>>> 1103587d92a4f715bf7988ac1a4b5b855e5ddf12
